requirements.txt
streamlit
sentence-transformers
faiss-cpu
pandas
torch
groq
python-dotenv
nest_asyncio
transformers